+++ src/godoctopus/src/levels/demo/DemoLevel.cpp
@@ -72,7 +72,7 @@ std::list<Steppable *> DemoLevelSteps(
 {
 	// Tick rate for anchor decay : easy => 1 every 2sec
 	// >= medium : 1 every sec
-	size_t anchorTickRate_l = difficulty_p < 2 ? 200 : 100;
+	uint32_t anchorTickRate_l = difficulty_p < 2 ? 200 : 100;
 
 	// number of spawn points
 	unsigned long nSpawnPoints_l = difficulty_p < 2 ? 1 : 2;
@@ -114,7 +114,7 @@ std::list<Steppable *> DemoLevelSteps(
 	}
 
 	std::list<WaveParam> params_l;
-	for(size_t i = 0 ; i < waves_l.size() ; ++ i)
+	for(uint32_t i = 0 ; i < waves_l.size() ; ++ i)
 	{
 		params_l.push_back({
 			{octopus::Vector(10,10), octopus::Vector(240,10), octopus::Vector(10,240), octopus::Vector(240,240)},
@@ -170,7 +170,7 @@ std::list<Steppable *> DemoLevelSteps(
 }
 
 AreaSpawnerCommand * createResourceNodeSpawnCommmand(Library &lib_p, RandomGenerator &rand_p, unsigned long x, unsigned long y, unsigned long size,
-	size_t nbRes_p, size_t nbAnchorSpot_p, size_t nbUnits_p, size_t qtyRes_p, size_t qtyIrium_p, size_t nbDoubleSquares_p, size_t nbQuadSquares_p, size_t nbUSquares_p)
+	uint32_t nbRes_p, uint32_t nbAnchorSpot_p, uint32_t nbUnits_p, uint32_t qtyRes_p, uint32_t qtyIrium_p, uint32_t nbDoubleSquares_p, uint32_t nbQuadSquares_p, uint32_t nbUSquares_p)
 {
 	std::list<AreaSpawn> spawners_l;
 
@@ -254,8 +254,8 @@ AreaSpawnerCommand * createResourceNodeSpawnCommmand(Library &lib_p, RandomGener
 std::list<Command *> DemoLevelCommands(Library &lib_p, RandomGenerator &rand_p, unsigned long playerCount_p, int difficulty_p)
 {
 	// resource on basic resources node based on difficulty
-	size_t basicRes_l = difficulty_p < 3 ? 4000 : 2000;
-	size_t advancedRes_l = difficulty_p < 3 ? 1000 : 500;
+	uint32_t basicRes_l = difficulty_p < 3 ? 4000 : 2000;
+	uint32_t advancedRes_l = difficulty_p < 3 ? 1000 : 500;
 
 	// createResourceNodeSpawnCommmand(lib, rand, x, y, size, nb res, nb anchor, nb units, qty irium)
 	std::list<Command *> commands_l {
@@ -293,7 +293,7 @@ std::list<Command *> DemoLevelCommands(Library &lib_p, RandomGenerator &rand_p, 
 void writeWaveContentInfo(std::ofstream &file_p, WaveContentInfo const &info_p)
 {
 	file_p.write((char*)&info_p.steps, sizeof(info_p.steps));
-	size_t size_l = info_p.units.size();
+	uint32_t size_l = info_p.units.size();
 	file_p.write((char*)&size_l, sizeof(size_l));
 	for(WaveUnitCount const &unitCount_l : info_p.units)
 	{
@@ -305,9 +305,9 @@ void writeWaveContentInfo(std::ofstream &file_p, WaveContentInfo const &info_p)
 void readWaveContentInfo(std::ifstream &file_p, WaveContentInfo &info_p)
 {
 	file_p.read((char*)&info_p.steps, sizeof(info_p.steps));
-	size_t size_l = 0;
+	uint32_t size_l = 0;
 	file_p.read((char*)&size_l, sizeof(size_l));
-	for(size_t i = 0 ; i < size_l ; ++ i)
+	for(uint32_t i = 0 ; i < size_l ; ++ i)
 	{
 		info_p.units.push_back(WaveUnitCount());
 		info_p.units.back().model = readString(file_p);
@@ -318,7 +318,7 @@ void readWaveContentInfo(std::ifstream &file_p, WaveContentInfo &info_p)
 void writeWavePoolInfo(std::ofstream &file_p, WavePoolInfo const &info_p)
 {
 	// write wave info
-	size_t size_l = info_p.infos.size();
+	uint32_t size_l = info_p.infos.size();
 	file_p.write((char*)&size_l, sizeof(size_l));
 	for(WaveInfo const &waveInfo_l : info_p.infos)
 	{
@@ -340,15 +340,15 @@ void writeWavePoolInfo(std::ofstream &file_p, WavePoolInfo const &info_p)
 void readWavePoolInfo(std::ifstream &file_p, WavePoolInfo &info_p)
 {
 	// read wave info
-	size_t size_l = 0;
+	uint32_t size_l = 0;
 	file_p.read((char*)&size_l, sizeof(size_l));
-	for(size_t i = 0 ; i < size_l ; ++ i)
+	for(uint32_t i = 0 ; i < size_l ; ++ i)
 	{
 		WaveInfo waveInfo_l;
 		// read upgrades
-		size_t sizeUpgrades_l = 0;
+		uint32_t sizeUpgrades_l = 0;
 		file_p.read((char*)&sizeUpgrades_l, sizeof(sizeUpgrades_l));
-		for(size_t j = 0 ; j < sizeUpgrades_l ; ++ j)
+		for(uint32_t j = 0 ; j < sizeUpgrades_l ; ++ j)
 		{
 			std::string up_l = readString(file_p);
 			waveInfo_l.upgrades.push_back(up_l);
@@ -370,7 +370,7 @@ void writeDemoLevelHeader(std::ofstream &file_p, DemoLevelHeader const &header_p
 	file_p.write((char*)&header_p.difficulty, sizeof(header_p.difficulty));
 	file_p.write((char*)&header_p.player_count, sizeof(header_p.player_count));
 
-	size_t size_l = header_p.tierWaveInfo.size();
+	uint32_t size_l = header_p.tierWaveInfo.size();
 	file_p.write((char*)&size_l, sizeof(size_l));
 
 	for(WavePoolInfo const &info_l : header_p.tierWaveInfo)
@@ -386,10 +386,10 @@ void readDemoLevelHeader(std::ifstream &file_p, DemoLevelHeader &header_r)
 	file_p.read((char*)&header_r.difficulty, sizeof(header_r.difficulty));
 	file_p.read((char*)&header_r.player_count, sizeof(header_r.player_count));
 
-	size_t size_l = 0;
+	uint32_t size_l = 0;
 	file_p.read((char*)&size_l, sizeof(size_l));
 
-	for(size_t i = 0 ; i < size_l ; ++ i)
+	for(uint32_t i = 0 ; i < size_l ; ++ i)
 	{
 		WavePoolInfo info_l;
 		readWavePoolInfo(file_p, info_l);
